requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-redis==5.0.0
django-simpleui==2021.6.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
